Move "printPage()" to `helpers.js` for reuse in multiple pages
