+++ client/homebrew/navbar/print.navitem.jsx
@@ -1,8 +1,9 @@
 const React = require('react');
 const Nav = require('naturalcrit/nav/nav.jsx');
+const { printPage } = require('../../../shared/helpers.js');
 
-module.exports = function({printFunction}){
-	return <Nav.item onClick={printFunction} color='purple' icon='far fa-file-pdf'>
+module.exports = function(){
+	return <Nav.item onClick={printPage} color='purple' icon='far fa-file-pdf'>
 		get PDF
 	</Nav.item>;
 };
