+++ client/homebrew/pages/sharePage/sharePage.jsx
@@ -9,11 +9,10 @@ const MetadataNav = require('../../navbar/metadata.navitem.jsx');
 const PrintNavItem = require('../../navbar/print.navitem.jsx');
 const RecentNavItem = require('../../navbar/recent.navitem.jsx').both;
 const Account = require('../../navbar/account.navitem.jsx');
-
-
 const BrewRenderer = require('../../brewRenderer/brewRenderer.jsx');
 
 const { DEFAULT_BREW_LOAD } = require('../../../../server/brewDefaults.js');
+const { printPage } = require('../../../../shared/helpers.js');
 
 const SharePage = createClass({
 	displayName     : 'SharePage',
@@ -35,20 +34,9 @@ const SharePage = createClass({
 		if(!(e.ctrlKey || e.metaKey)) return;
 		const P_KEY = 80;
 		if(e.keyCode == P_KEY){
-			if(e.keyCode == P_KEY) this.printPage();
+			if(e.keyCode == P_KEY) printPage();
 			e.stopPropagation();
 			e.preventDefault();
-		}
-	},
-
-	printPage : function(){
-		if (window.typeof !== 'undefined') {
-			window.frames['BrewRenderer'].contentWindow.print();
-			//Force DOM reflow; Print dialog causes a repaint, and @media print CSS somehow makes out-of-view pages disappear
-			let node = window.frames['BrewRenderer'].contentDocument.getElementsByClassName('brewRenderer').item(0);
-			node.style.display='none';
-			node.offsetHeight; // accessing this is enough to trigger a reflow
-			node.style.display='';
 		}
 	},
 
@@ -83,7 +71,7 @@ const SharePage = createClass({
 
 				<Nav.section>
 					{this.props.brew.shareId && <>
-						<PrintNavItem printFunction={this.printPage}/>
+						<PrintNavItem/>
 						<Nav.dropdown>
 							<Nav.item color='red' icon='fas fa-code'>
 								source
